Load icon data asynchronously instead of using thread pool
This does leak the memory; it is assumed that once an application is in polymodo's list, it'll persist there forever. That is of course not the case, but it suffices for 99.9% of launches; and at most you'll have one or two icons that aren't used anymore but remain in memory.